+++ src/Narvalo.Web/UI/Assets/AssetProviderBase.cs
@@ -33,6 +33,7 @@ namespace Narvalo.Web.UI.Assets
         {
             Contract.Ensures(Contract.Result<Uri>() != null);
 
+            // REVIEW: Seems odd to return null when we require the opposite.
             return default(Uri);
         }
 
@@ -40,6 +41,7 @@ namespace Narvalo.Web.UI.Assets
         {
             Contract.Ensures(Contract.Result<Uri>() != null);
 
+            // REVIEW: Seems odd to return null when we require the opposite.
             return default(Uri);
         }
 
@@ -47,6 +49,7 @@ namespace Narvalo.Web.UI.Assets
         {
             Contract.Ensures(Contract.Result<Uri>() != null);
 
+            // REVIEW: Seems odd to return null when we require the opposite.
             return default(Uri);
         }
 
@@ -54,6 +57,7 @@ namespace Narvalo.Web.UI.Assets
         {
             Contract.Ensures(Contract.Result<Uri>() != null);
 
+            // REVIEW: Seems odd to return null when we require the opposite.
             return default(Uri);
         }
     }